+++ src/CTBX.Backend/Uploads/sample_holidays_file2.csv
@@ -1,0 +1,3 @@
+USA;California;Independence Day;2023-07-04;True
+USA;Texas;Thanksgiving;2023-11-23;True
+Canada;Quebec;National Holiday;2023-06-24;False
