+++ src/CTBX.ImportHoliday.UI/HolidayView.razor.cs
@@ -1,63 +1,71 @@
-﻿//using CTBX.CommonMudComponents;
-//using CTBX.EmployeesImport.Shared;
-//using CTBX.ImportHoliday.UI;
-//using CTBX.ImportHolidays.Shared;
-//using Microsoft.AspNetCore.Components;
-//using MudBlazor;
-//using static MudBlazor.CategoryTypes;
+﻿using System.Collections.Immutable;
+using CTBX.CommonMudComponents;
+using CTBX.EmployeesImport.Shared;
+using CTBX.ImportHoliday.UI;
+using CTBX.ImportHolidays.Shared;
+using Microsoft.AspNetCore.Components;
+using MudBlazor;
+using static MudBlazor.CategoryTypes;
 
 
-//namespace CTBX.ImportHolidays.UI
-//{
-//    public class HolidayViewBase : BaseMudComponent
-//    {
-//        [Inject]
-//        public required ImportHolidaysService Service { get; set; }
-//        public List<Holiday> HolidaysList = new();
-//        public string _searchString = string.Empty;
-//        public bool _sortNameByLength;
+namespace CTBX.ImportHolidays.UI
+{
+    public class HolidayViewBase : BaseMudComponent
+    {
+        [Inject]
+        public required ImportHolidaysService Service { get; set; }
+        protected IImmutableList<Holiday> HolidaysList { get; set; } = ImmutableList<Holiday>.Empty;
 
-//        protected override async Task OnInitializedAsync()
-//        {
-//            await ReloadData();
-//        }
+        public string _searchString = string.Empty;
+        public bool _sortNameByLength;
 
-//        public Func<Holiday, object> _sortBy => x =>
-//        {
-//            if (_sortNameByLength)
-//                return x.Country;
-//            else
-//                return x.HolidayDate;
-//        };
+        protected override async Task OnInitializedAsync()
+        {
+            await ReloadHolidaysData();
+        }
 
-//        public Func<Holiday, bool> _quickFilter => x =>
-//        {
-            
-//            if (string.IsNullOrWhiteSpace(_searchString))
-//                return true;
+        public Func<Holiday, object> _sortBy => x =>
+        {
+            if (_sortNameByLength)
+                return x.Country;
+            else
+                return x.HolidayDate;
+        };
 
-//            if (x.Country.Contains(_searchString, StringComparison.OrdinalIgnoreCase))
-//                return true;
+        public Func<Holiday, bool> _quickFilter => x =>
+        {
 
-//            if (x.HolidayDate.ToString().Contains(_searchString, StringComparison.OrdinalIgnoreCase))
-//                return true;
+            if (string.IsNullOrWhiteSpace(_searchString))
+                return true;
 
-//            if (x.HolidayName.Contains(_searchString, StringComparison.OrdinalIgnoreCase))
-//                return true;
+            if (x.Country.Contains(_searchString, StringComparison.OrdinalIgnoreCase))
+                return true;
 
-//            if (x.IsGlobal.ToString().Contains(_searchString, StringComparison.OrdinalIgnoreCase))
-//                return true;
+            if (x.HolidayDate.ToString().Contains(_searchString, StringComparison.OrdinalIgnoreCase))
+                return true;
 
-        
-//            return false;
-//        };
+            if (x.HolidayName.Contains(_searchString, StringComparison.OrdinalIgnoreCase))
+                return true;
 
-//        public async Task ReloadData()
-//        {
-            
-//            var result = await service.getholidaysasync();
-//            holidayslist = [.. result];
-//        }
+            if (x.IsGlobal.ToString().Contains(_searchString, StringComparison.OrdinalIgnoreCase))
+                return true;
 
-//    }
-//}
+
+            return false;
+        };
+
+        public async Task ReloadHolidaysData()
+        {
+            try
+            {
+                HolidaysList = await Service.GetHolidaysAsync();
+
+            }
+            catch (Exception ex)
+            {
+                await NotifyError($"Failed to reload Holidays: {ex.Message}");
+                HolidaysList = ImmutableList<Holiday>.Empty; // Reset the list to an empty state
+            }
+        }
+    }
+}
