+++ src/CTBX.ImportHolidays.Shared/Holiday.cs
@@ -6,6 +6,6 @@ public class Holiday
     public string Country { get; set; } = string.Empty;
     public string State { get; set; } = string.Empty;
     public string HolidayName { get; set; } = string.Empty;
-    public DateOnly HolidayDate { get; set; } 
+    public DateTime HolidayDate { get; set; } 
     public bool IsGlobal { get; set; }
 }
