+++ src/CTBX.WebPortal/CTBX.WebPortal.Client/Layout/NavMenu.razor
@@ -14,6 +14,7 @@
 
         <MudNavLink Href="/employees-view" Match="NavLinkMatch.Prefix">My Employees</MudNavLink>
 		<MudNavLink Href="/holidays-upload" Match="NavLinkMatch.Prefix">Holidays Import</MudNavLink>
+		<MudNavLink Href="/holidays-view" Match="NavLinkMatch.Prefix">Holidays</MudNavLink>
     </MudNavGroup>
 
     <MudNavGroup Title="Future Features 2" Icon="@Icons.Material.Filled.Notes" Expanded="true">
